requirements.txt
Speedtest-cli